refactor(client): remove redundant code
