+++ src/client/constants.ts
@@ -2,13 +2,8 @@ import { type NativeFeaturesFromVersion } from './types';
 
 export const ANDROID_APP_ID = 'alfabank';
 
-export const COOKIE_KEY_BRIDGE_TO_NATIVE_NATIVE_HISTORY_STACK = 'bridgeToNativeData';
-
 export const DEEP_LINK_PATTERN =
     /^(\/|alfabank:\/{3}dashboard\/|alfabank:\/{3}|alfabank:\/{2}|https:\/{2}online.alfabank.ru\/)/;
-
-export const QUERY_CLOSE_WEBVIEW_KEY = 'closeWebView';
-export const QUERY_CLOSE_WEBVIEW_VALUE = 'true';
 
 export const NATIVE_FEATURES_FROM_VERSION: NativeFeaturesFromVersion = {
     android: {
@@ -26,10 +21,6 @@ export const NATIVE_FEATURES_FROM_VERSION: NativeFeaturesFromVersion = {
         savedBackStack: { fromVersion: '0.0.0' },
     },
 } as const;
-
-export const PREVIOUS_B2N_STATE_STORAGE_KEY = 'previousBridgeToNativeState';
-export const PREVIOUS_NATIVE_NAVIGATION_AND_TITLE_STATE_STORAGE_KEY =
-    'previousNativeNavigationAndTitleState';
 
 export const VERSION_TO_IOS_APP_ID = {
     '0.0.0': 'alfabank',
