--- src/server/constants.ts
@@ -1,33 +1,0 @@
-/*
- * Ключи стандартных заголовков.
- */
-
-export const HEADER_KEY_COOKIE = 'cookie';
-export const HEADER_KEY_USER_AGENT = 'user-agent';
-
-/*
- * Ключи заголовков и query-параметров, которые использует нативное приложение,
- * для передачи информации о себе в веб.
- */
-
-// Нативное приложение на обеих платформах подмешивает этот заголовок к запросу за HTML.
-// TODO:
-// * Исследовать, делает ли оно это при запросах к прочим ресурсам;
-// * Есть подозрение на некоторую нестабильность — возможно делает это
-//  только для первого документа в вебвью-сессии. Исследовать.
-export const HEADER_KEY_NATIVE_APPVERSION = 'app-version';
-
-// Нативное iOS приложение передаёт в этом параметре схему,
-// под которым оно зарегистрировано в OS.
-// Известные проблемы:
-// * В старых версиях отсутствует, клиентский код использует хардкод (см. `src/constants.ts`);
-// * В тестовых сборках часто нарушена связь!!! Между фактической схемой
-//  и схемой, которая приходит в этом параметре.
-//  Нужно просить iOS разработчика собрать сборку нормально.
-export const QUERY_NATIVE_IOS_APPID = 'applicationId';
-
-// Нативное iOS приложение передаёт в этом параметре свою версию.
-export const QUERY_NATIVE_IOS_APPVERSION = 'device_app_version';
-
-// Нативное приложение на обеих платформах в этом параметре передаёт активную тему (светлая/тёмная).
-export const QUERY_NATIVE_THEME = 'theme';
