+++ src/server/is-webview-env.ts
@@ -1,4 +1,5 @@
-import { HEADER_KEY_NATIVE_APPVERSION, HEADER_KEY_USER_AGENT } from './constants';
+import { HEADER_KEY_NATIVE_APPVERSION, HEADER_KEY_USER_AGENT } from '../query-and-headers-keys';
+
 import { versionPattern, webviewUaIOSPattern } from './regexp-patterns';
 import { type UniversalRequest } from './types';
 import { getHeaderValue, hasBridgeToNativeDataCookie } from './utils';
