+++ src/server/prepare-native-app-details-for-client.ts
@@ -1,17 +1,15 @@
 import {
     COOKIE_KEY_BRIDGE_TO_NATIVE_DATA,
-    QUERY_B2N_TITLE,
-    QUERY_B2N_TITLE_DEPRECATED,
-} from '../constants';
-import { type NativeParams } from '../types';
-
-import {
     HEADER_KEY_NATIVE_APPVERSION,
     QUERY_B2N_NEXT_PAGEID,
+    QUERY_B2N_TITLE,
+    QUERY_B2N_TITLE_DEPRECATED,
     QUERY_NATIVE_IOS_APPID,
     QUERY_NATIVE_IOS_APPVERSION,
     QUERY_NATIVE_THEME,
-} from './constants';
+} from '../query-and-headers-keys';
+import { type NativeParams } from '../types';
+
 import { extractNativeServiceQueries } from './extract-native-service-queries';
 import { iosAppIdPattern, versionPattern } from './regexp-patterns';
 import { type UniversalRequest } from './types';
